blog/requirements.txt
appnope==0.1.0
decorator==4.2.1
Django==2.0.1
django-filter==1.1.0
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
mongoengine==0.15.0
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
PyJWT==1.6.0
pymongo==3.6.1
python-dotenv==0.8.2
pytz==2018.3
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
